[WCNCR00262566] wlan: refacotr pointer casting for calculation (2/.)
[Description]
refacotr pointer casting for calculation

MTK-Commit-Id: ad94f99039a52a6ee93e4b7529141abe8d3a5cdf

CR-Id: WCNCR00262566
Feature: wlan/os
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/6096477
CheckCourse: srv_pis <srv_pis@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>
Tested-by: yuchi.wang <yuchi.wang@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Test: srv_mspautosanity

+++ mgmt/assoc.c
@@ -683,13 +683,11 @@ struct MSDU_INFO* assocComposeReAssocReqFrame(IN struct ADAPTER *prAdapter,
 
 	/* Compose Header and Fixed Field */
 	assocComposeReAssocReqFrameHeaderAndFF(prAdapter,
-					       prStaRec,
-					       (uint8_t
-						*) ((unsigned
-						     long)(prMsduInfo->prPacket)
-						    + MAC_TX_RESERVED_FIELD),
-					       prBssInfo->aucOwnMacAddr,
-					       &u2PayloadLen);
+			prStaRec,
+			(uint8_t *)((uintptr_t)(prMsduInfo->prPacket)
+			+ MAC_TX_RESERVED_FIELD),
+			prBssInfo->aucOwnMacAddr,
+			&u2PayloadLen);
 
 	/* 4 <3> Update information of MSDU_INFO_T */
 	nicTxSetPktLifeTime(prAdapter, prMsduInfo, 100);
